webapi/tct-animations-css3-tests/animations/CSS3Animations_animation_shorthand.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu, Jinfeng <jinfengx.liu@intel.com>
        Li, Hao <haox.li@intel.com>
-->

<html>
  <head>
    <title>CSS3 Animations Test: CSS3Animations_animation_shorthand</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-css3-animations-20120403/#animation" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if animation shorthand property combines six of the animation properties into a single property" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="support/support.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
    <script>
        test(function () {
            var div = document.querySelector("#test");
            div.style[headProp("animation")] = "sample 5s linear 5s 2 alternate none";
            var animationName = div.style.getPropertyValue("animation-name") || div.style.getPropertyValue("-webkit-animation-name");
            assert_equals(animationName, "sample", "The element animation-name value Test");
            var animationDuration = div.style.getPropertyValue("animation-duration") || div.style.getPropertyValue("-webkit-animation-duration");
            assert_equals(animationDuration, "5s", "The element animation-duration value Test");
            var animationTimingFunction = div.style.getPropertyValue("animation-timing-function") || div.style.getPropertyValue("-webkit-animation-timing-function");
            assert_equals(animationTimingFunction, "linear", "The element animation-timing-function value Test");
            var animationDelay = div.style.getPropertyValue("animation-delay") || div.style.getPropertyValue("-webkit-animation-delay");
            assert_equals(animationDelay, "5s", "The element animation-delay value Test");
            var animationIterationCount = div.style.getPropertyValue("animation-iteration-count") || div.style.getPropertyValue("-webkit-animation-iteration-count");
            assert_equals(animationIterationCount, "2", "The animation-iteration-count value Test");
            var animationDirection = div.style.getPropertyValue("animation-direction") || div.style.getPropertyValue("-webkit-animation-direction");
            assert_equals(animationDirection, "alternate", "The animation-direction value Test");
            var animationFillMode = div.style.getPropertyValue("animation-fill-mode") || div.style.getPropertyValue("-webkit-animation-fill-mode");
            assert_equals(animationFillMode, "none", "The animation-fill-mode value Test");
        }, document.title);
    </script>
  </body>
</html>
